+++ grpc/config.go
@@ -1,5 +1,7 @@
 package grpc
 
+import "google.golang.org/grpc"
+
 // Config holds a list of options used during the grpc server construction.
 type Config struct {
 	Network string
@@ -7,6 +9,9 @@ type Config struct {
 
 	// Register the reflection API or not
 	ReflectionEnabled bool
+
+	// A list of arbitrary server options for the gRPC server
+	Options []grpc.ServerOption
 }
 
 // NewConfig returns a new config populated with default values.